JBIDE-21447 - Create OpenShift Server Adapter from service/route context menu
Refactored the ServerSettingsWizardFragment to extract the widgets
 into ServerSettingsWizardPage and keep the WizardFragment that now
 includes the WizardPage via a wrapper.
Added context menu entries in the OpenShift Explorer view and in the
 Properties view
When opening the wizard, the Eclipse project, the OpenShift service and
 route are set from the selected service/route (if match found)

Exported the "org.jboss.tools.openshift.core.util" package to be able to use
 it in the org.jboss.tools.openshift.ui project.
