+++ plugins/org.jboss.tools.openshift.client/.classpath
@@ -2,7 +2,7 @@
 <classpath>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry exported="true" kind="lib" path="lib/openshift-restclient-java-4.0.0-SNAPSHOT.jar"/>
+        <classpathentry exported="true" kind="lib" path="lib/openshift-restclient-java-4.0.0-SNAPSHOT.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/commons-compress-1.8.1.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/jboss-dmr-1.3.0.Final.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/log4j-1.2.17.jar"/>
