+++ plugins/org.jboss.tools.openshift.ui/plugin.properties
@@ -17,6 +17,7 @@ openshift.command.showproperties = Properties
 openshift.command.pod.logs = Pod Log...
 openshift.command.project.manage = Manage Projects...
 openshift.command.editresource=Edit...
+openshift.command.serverAdapter=Server Adapter...
 openshift.preferences.page.title = OpenShift 3
 openshift.preferences.page.certificates.title=SSL certificates
